phase-6.11: align piecewise out_cache_loc_swa dtype with regular runner
The shared `_forward_input_buffer_pool` (input_buffers.py:24-26) asserts
the same buffer name has the same dtype across all registrations. Phase
5.5 fixed cuda_graph_runner.py:186 to allocate `out_cache_loc_swa` as
`torch.int32` (matching `translate_loc_from_full_to_swa`'s return
dtype at swa_memory_pool.py:154). The piecewise runner still allocated
the same buffer as `torch.int64`, tripping the dtype-mismatch assert
during graph capture finalization.

Found via journal traceback at `init_piecewise_cuda_graphs` →
`PiecewiseCudaGraphRunner.__init__` → `buffers.share_buffers()` →
`AssertionError: Buffer out_cache_loc_swa has different dtype than before`.

+++ python/sglang/srt/model_executor/piecewise_cuda_graph_runner.py
@@ -217,7 +217,13 @@ class PiecewiseCudaGraphRunner:
                 (self.max_num_tokens,), dtype=self._cache_loc_dtype()
             )
             out_cache_loc_swa = (
-                torch.zeros((self.max_num_tokens,), dtype=torch.int64)
+                # int32 to match `translate_loc_from_full_to_swa`'s return
+                # dtype (swa_memory_pool.py:154) and the regular CUDA graph
+                # runner's allocation at cuda_graph_runner.py:186 (fixed in
+                # Phase 5.5). The shared `_forward_input_buffer_pool` asserts
+                # consistent dtype between the two graph runners, so they
+                # must agree.
+                torch.zeros((self.max_num_tokens,), dtype=torch.int32)
                 if model_runner.is_hybrid_swa
                 else None
             )
